Reduce the GPS coordinate decimal accuracy to 5 decimals, this will give us 1 meter accuracy which is enough, and it will make the matching of existing coordinates more likely

+++ service/src/main/java/fi/poltsi/vempain/file/repository/GpsLocationRepository.java
@@ -4,9 +4,11 @@ import fi.poltsi.vempain.file.entity.GpsLocationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
 public interface GpsLocationRepository extends JpaRepository<GpsLocationEntity, Long> {
-	Optional<GpsLocationEntity> findByLatitudeAndLatitudeRefAndLongitudeAndLongitudeRef(Double latitude, Character latitudeRef, Double longitude, Character longitudeRef);
+	Optional<GpsLocationEntity> findByLatitudeAndLatitudeRefAndLongitudeAndLongitudeRef(BigDecimal latitude, Character latitudeRef, BigDecimal longitude,
+																						Character longitudeRef);
 }
